op-deployer: Add deploy mips script (#12835)
* op-deployer: Add deploy mips script

* op-deployer: Explicitly set the PreimageOracle to use with deployed MIPS

* op-deployer: Validate mips version in go code.

+++ op-deployer/pkg/deployer/opcm/dispute_game.go
@@ -11,7 +11,7 @@ import (
 type DeployDisputeGameInput struct {
 	Release                  string
 	StandardVersionsToml     string
-	MipsVersion              uint8
+	MipsVersion              uint64
 	MinProposalSizeBytes     uint64
 	ChallengePeriodSeconds   uint64
 	GameKind                 string
